Update build-system requirement for setuptools-scm to >=7.0.0
pyproject.toml:
Update build-system requirement for setuptools-scm to >=7.0.0, which
obsoletes the use of setuptools-scm-git-archive.
Remove pip and wheel from build-system requirements, as they are not
needed in a PEP517 workflow facilitating pypa/build and pypa/installer.

setup.{cfg,py}:
Update build-system requirement for setuptools-scm to >=7.0.0, which
obsoletes the use of setuptools-scm-git-archive.