Make `test_run_string` pass when the default shell is not bash
My default shell is fish (https://fishshell.com).
`test_run_string` failed on my laptop because the `cmd` was not valid fish syntax.

One option is to find a command which is similar to what was there but works in bash and other shells.
Another option, the one I went with, was to always call bash.